Plumbing planIds through the client
This change allows the web client code to fetch plan ids from the server, and then pass them back to the client.
It retains the existing behavior of the current examples while being a little more round about, and opening the door for work in each server to consume planIds from the client.